requirements.txt
telethon==1.36.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
